Move scrap movies data to separate file
