+++ components/web-scrapping/scrap-movies-data.spec.js
@@ -1,0 +1,11 @@
+const scrapMoviesData = require('./scrap-movies-data');
+
+describe('Scrap Movies Data', () => {
+  it('should be defined', () => {
+    expect(scrapMoviesData).toBeDefined();
+  });
+
+  it('should be function', () => {
+    expect(typeof scrapMoviesData).toEqual('function');
+  });
+});
